Caching, file types, composer, cleaner metadata format
On-demand caching as been added. When a page is requested, a cached
version will be checked. If it exists and the current hash of its
source matches the hash when the cached page was created, it will be
displayed. Otherwise the page will be generated as usual and a new
cached version will be created. MD5 hashes are used to determine if
a cached needs to be renewed.

The metadata format at the top of pages and articles has been slightly
changed to the format:

---
metadata
---
content

This allows for a clean display of metadata, and more fault tolerant
to parse. A new metadata tag has been added "TYPE". This can be set
if different from the set default type. This tag is used to determine
how the content should be displayed. This allows for mixed pages of
markdown and plain text and even HTML.

Composer is now used for the Parsedown library. Version 1.5.2 has
been selected.

Overall, the code has been separated into functions. The template
has been separated and thined out. The article list is no longer
tied to a specific page, anywhere the tag {{article list}} is placed
in a Markdown document, it will be replaced with the article list.